Add protocol 774 (MC 1.21.11) support
Multi-protocol support for 772 (1.21.8) and 774 (1.21.11) with
auto-detection via server status query.

Protocol changes:
- VarInt entity IDs in 774 (was VarLong), LpVec3 velocity
- Fix VarInt encoding of negative values (was infinite loop)
- Data component registry keyed by name, per protocol version
- PickItem replaced with block-position-based PickItemFromBlock
- PlayTransfer packet for PLAY-state server transfers
- TypedEntityData for bucket/block/entity_data in 774
- AcknowledgeBlockChange simplified to sequence ID (post-1.19)
- UnloadChunk encoding fixed to ChunkPos format

Gameplay mechanics:
- Sprint state changes sent via EntityAction packets
- Slot state_id echoes server value instead of incrementing
- Block breaking speed: tool component and material fallback
- Heightmap type IDs corrected to match vanilla enum order

Architecture:
- MCData version-aware loading (protocol-based selection)
- SpectateServer rewritten for multi-protocol support
- Simplified ConfigurationDisconnect and chat NBT parsing
- Deduplicated TRANSLATIONS constant across Chat/TextComponent

Testing and CI:
- Docker compose supports MC_VERSION env for matrix testing
- Reset protocol_version in unit specs to prevent CI poisoning
- Fix ameba linter warnings for bool property naming

+++ src/rosegold/models/entity_effect.cr
@@ -50,7 +50,7 @@ class Rosegold::EntityEffect
 
   property \
     effect : Effect,
-    amplifier : UInt8,
+    amplifier : UInt32,
     duration : UInt32,
     flags : UInt8
